unsigned short --> int: unsigned short было нужно для приведения к типу Паскаля word в 18 задаче. В этой задаче лучше использовать int

+++ 39/prog.cc
@@ -3,32 +3,33 @@
 #include <algorithm> 
 using namespace std;
 
-void outputArray(short* array, unsigned short& size) {
-    cout << " array = [ ";
-    for(unsigned short i = 0; i < size; i++)
+void outputArray(int* array, int& size) {
+  cout << " array = [ ";
+  for(int i = 0; i < size; i++)
     cout << array[i] << ", "; 
-    cout << "]" << endl;
+  cout << "]" << endl;
 }
 
 int main(int argc, char const *argv[]){
-    unsigned short size = 3;
-    cout << "Input size of array = "; cin >> size;
-    short* array = new short[size];
-    cout << "Input array:" << endl;
-    for(unsigned short i = 0; i < size; i++) {
+  int size = 3;
+  cout << "Input size of array = "; cin >> size;
+  int* array = new int[size];
+
+  cout << "Input array:" << endl;
+  for(int i = 0; i < size; i++) {
     cout << " array[" << i << "] = "; cin >> array[i];
-    }
+  }
 
-    cout << "\nInputed array:" << endl;
-    outputArray(array, size);
-	short n = 5; cout << "Input N = "; cin >> n;
+  cout << "\nInputed array:" << endl;
+  outputArray(array, size);
+	int n = 5; cout << "Input N = "; cin >> n;
 
-	unsigned short count = 0;
-	for(unsigned short i = 0; i < size; i++)
+	int count = 0;
+	for(int i = 0; i < size; i++)
 		count += array[i] == n ? 1 : 0;
 
 	cout << "\nN occurrences in inputed array " << count << " times " << endl;
 
-    return 0;
+  return 0;
 }
